rework partials, partl_cb vs idx
