+++ src/Deprecated.jl
@@ -1,5 +1,17 @@
 
 
+# @kwdef struct MvNormalKernel{P, T, M, iM} <: AbstractKernel
+#     """ On-manifold point representing center (mean) of the MvNormal distribution """
+#     μ::P
+#     """ Zero-mean normal distribution with covariance """
+#     p::MvNormal{T, M}
+#     # TDB might already be covered in p.Σ.chol but having issues with SymPD (not particular to this AMP repo)
+#     """ Manually maintained square root concentration matrix for faster compute, TODO likely duplicate of existing Distrubtions.jl functionality. """
+#     sqrt_iΣ::iM = sqrt(inv(cov(p)))
+#     """ Nonparametric weight value """
+#     weight::Float64 = 1.0
+# end
+
 
 # # deprecated as legacy and replaced by previously called manikde!_manellic
 # function manikde!(
