+++ src/ExportAPI.jl
@@ -34,3 +34,7 @@ export mmd!, mmd
 
 # partial specific functions
 export getManifoldPartial
+
+
+## TODO, export
+# export DensityKernel, ConcentratedGaussianKernel
